requirements.txt
PySide6>=6.7.0
openai>=1.40.0
edge-tts>=6.1.12
requests>=2.31.0
python-dotenv>=1.0.1
